requirements.txt
Flask==0.10.1
Flask-Cache==0.13.1
Pillow==2.4.0
